pyproject.toml
[tool.poetry]
name = "mlsgpt"
version = "0.1.0"
description = ""
authors = ["Kwesi P Apponsah <kwesiparry@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
mlsgpt = "mlsgpt.cmd:cli"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
redis = "^5.0.3"
pymupdf = "^1.24.1"
python-dotenv = "^1.0.1"
click = "^8.1.7"
httpx = "^0.27.0"
requests = "^2.31.0"
lancedb = "^0.6.8"
pgvector = "^0.2.5"
openai = "^1.17.1"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
psycopg = "^3.1.18"
uvicorn = "^0.29.0"
pyngrok = "^7.1.6"
jinja2 = "^3.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
kombu = "^5.3.7"
boto3 = "^1.34.88"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
pre-commit = "^3.7.0"
flake8 = "^7.0.0"
mypy = "^1.9.0"
isort = "^5.13.2"
ruff = "^0.3.7"
black = {version = "^24.4.0", allow-prereleases = true}
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
